assets/requirements.txt
urllib3
sseclient-py
requests